Redefine User model: Settings already unmarshalled map[string]any
This bids for data validation when Settings are inserted into the
database.  These cannot be just any JSON, particularly an array, but
only a `map[string]any`-comliant objects.

With PostgreSQL (as of 16.1) JSON schema validation is not natively
supported.  There are plugins for that ([1], [2]), but I am not inclined
towards using them - makes setup slightly more complex.  Instead let's
make sure that only JSON objects are allowed to be stored as
`User.Settings` in Go functions which will insert/update these records.

[1] https://github.com/gavinwahl/postgres-json-schema
[2] https://github.com/supabase/pg_jsonschema